refactor: Profile 어노테이션 등록
- string으로 입력시 오타가 있을 수 있어 커스텀 어노테이션으로 등록
