+++ src/main/java/capstone/examlab/config/profile/ProfileMongo.java
@@ -1,0 +1,7 @@
+package capstone.examlab.config.profile;
+
+import org.springframework.context.annotation.Profile;
+
+@Profile("mongo")
+public @interface ProfileMongo {
+}
