+++ src/main/java/capstone/examlab/users/repository/UserMongoRepository.java
@@ -1,9 +1,10 @@
 package capstone.examlab.users.repository;
 
+import capstone.examlab.config.profile.ProfileMongo;
 import capstone.examlab.users.domain.UserDoc;
 import org.springframework.context.annotation.Profile;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
-@Profile("mongo")
+@ProfileMongo
 public interface UserMongoRepository extends MongoRepository<UserDoc, String>, UserRepository<UserDoc>{
 }
